+++ inst/NEWS.md
@@ -1,5 +1,5 @@
 
-# parsedate dev
+# parsedate 1.2.0
 
 * `parse_date()` and `parse_iso_8601()` now dupport a default time zone,
   that will be used for dates that do not explicitly specify one.